requirements.txt
streamlit
azure-ai-vision-imageanalysis
Pillow
